+++ ActivityScheduler.Shared/Service/CommonOperationResult.cs
@@ -8,17 +8,17 @@ namespace ActivityScheduler.Shared
 {
     public class CommonOperationResult
     {
-        public bool _success;
-        public string _msg;
-        public object _returningValue;
+        public bool Success { get; set; }
+        public string Message { get; set; }
+        public object ReturningValue { get; set; }
 
         public static CommonOperationResult GetInstance(bool success, string msg, object returningValue = null)
         {
             CommonOperationResult c = new CommonOperationResult()
             {
-                _msg = msg,
-                _returningValue = returningValue,
-                _success = success
+                Message = msg,
+                ReturningValue = returningValue,
+                Success = success
             };
             return c;
         }
@@ -27,6 +27,6 @@ namespace ActivityScheduler.Shared
         public static CommonOperationResult SayOk(string _msg = "") { return GetInstance(true, _msg, null); }
         public static CommonOperationResult SayItsNull(string _msg = "") { return GetInstance(true, _msg, null); }
 
-        public string ShrotString() => $"Success: {_success} message: {_msg}";
+        public string ShrotString() => $"Success: {Success} message: {Message}";
     }
 }