+++ MainWindow.xaml.cs
@@ -27,19 +27,21 @@ namespace ActivityScheduler
     {
         private ActivityScheduler.Core.Settings.Settings settingsFrm;
         private ActivitySchedulerApp _app;
+        WorkerServiceManager _workerMgr;
 
         SettingsManager _settingsManager;
-        public MainWindow(SettingsManager settingsManager, ActivitySchedulerApp app)
+        public MainWindow(SettingsManager settingsManager, ActivitySchedulerApp app, WorkerServiceManager workerMgr)
         {
             InitializeComponent();
             _settingsManager = settingsManager;
             _app = app;
+            _workerMgr = workerMgr;
             InitializeComponent();
         }
        
         private void Button_Click(object sender, RoutedEventArgs e)
         {
-            settingsFrm = new Core.Settings.Settings(_settingsManager, _app);
+            settingsFrm = new Core.Settings.Settings(_settingsManager, _app, _workerMgr);
             settingsFrm.ShowDialog();
         }
     }