+++ Releases.md
@@ -1,3 +1,9 @@
+## 1.1.0
+
+- #11: Create simple windows-service app that will start main
+- #12: The only one instance of application should be opened
+
+
 ## 1.0.0
 
 - #10: Create WPF app basis with logger and settings on EF-SQLITE